requirements.txt
affine==2.4.0
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
dask==2025.3.0
et_xmlfile==2.0.0
fonttools==4.57.0
fsspec==2025.3.2
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
msal==1.32.0
narwhals==1.35.0
numpy==2.2.4
Office365-REST-Python-Client==2.6.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
partd==1.4.2
pathlib==1.0.1
pdfkit==1.0.0
pillow==11.2.1
plotly==6.0.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pydyf==0.11.0
PyJWT==2.10.1
pyogrio==0.10.0
pyparsing==3.2.3
pyphen==0.17.2
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
shapely==2.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
streamlit-aggrid==1.1.2
template==0.7.6
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xarray==2025.3.1
XlsxWriter==3.2.2
zopfli==0.2.3.post1
